requirements.txt
# MedGenomics Backend Requirements
# Install with: pip install -r requirements.txt

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database Drivers
# Uncomment the one you're using:
sqlalchemy==2.0.23        # ORM
# pymongo==4.6.0            # MongoDB driver
psycopg2-binary==2.9.9   # PostgreSQL driver

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# API & CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Optional: For async support
aiofiles==23.2.1

# Optional: For WebSocket testing
websocket-client==1.6.4

# Optional: Database Migrations
alembic==1.13.0
